% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_oi_class.R
\docType{class}
\name{abun_oi-class}
\alias{abun_oi-class}
\title{abun_oi class}
\description{
A S4 class is used to present the data, model specification, and maximum empirical likelihood (EL) estimation results under one-inflated count regression models.
}
\section{Slots}{

\describe{
\item{\code{model}}{character specification of \code{"zt"} (zero-truncated model without one-inflation), \code{"ztoi"} (zero-truncated one-inflated model), or \code{"oizt"} (one-inflated zero-truncated model).}

\item{\code{dist}}{character specification of count regression model family, \code{"poisson"} or \code{"binomial"}.}

\item{\code{N}}{number representing the maximum EL estimate of abundance \eqn{N}.}

\item{\code{ci}}{vector representing the EL ratio confidence interval of abundance.}

\item{\code{beta}}{vector representing the maximum EL estimate of regression coefficients in count regression model.}

\item{\code{eta}}{vector representing the maximum EL estimate of regression coefficients in logistic regression model for one-inflation.}

\item{\code{alpha}}{number representing the maximum EL estimate of the probability of never being captured.}

\item{\code{loglikelihood}}{number representing the log-EL value.}

\item{\code{AIC}}{number representing the Akaike Information Criterion value.}

\item{\code{prob}}{vector representing the probability masses of individual covariates.}

\item{\code{nit}}{number representing the number of iterations of EM algorithm.}

\item{\code{pars_trace}}{matrix. Row shows the value of parameters in each iteration of EM algorithm.}

\item{\code{loglikelihood_trace}}{vector. Element represents the value of log-likelihood in each iteration of EM algorithm.}

\item{\code{y}}{vector specifying the number of times that individuals were captured.}

\item{\code{K}}{number specifying the number of capture occasions.}

\item{\code{x}}{matrix specifying the individual covariates in count regression model.}

\item{\code{z}}{matrix specifying the individual covariates in one-inflated regression model.}

\item{\code{epsilon}}{positive convergence tolerance \eqn{\epsilon}. The iterations converge in EM algorithm when the increase of the log-EL is less than \eqn{\epsilon}.}

\item{\code{maxN}}{number specifying the largest searching value for \eqn{N} in EM algorithm.}

\item{\code{maxN_ci}}{number specifying the largest searching value for calculating the confidence interval of \eqn{N}.}

\item{\code{level}}{number specifying the nominal level of confidence interval of \eqn{N}.}

\item{\code{maxit}}{integer specifying the maximal number of iterations in EM algorithm.}

\item{\code{start_beta}}{vector specifying the starting value of the coefficient \eqn{\beta} in count regression model in EM algorithm.}

\item{\code{start_eta}}{vector specifying the starting value of the coefficient \eqn{\eta} in one-inflated regression model in EM algorithm.}
}}

