% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_oi.R
\name{abun_oi.control}
\alias{abun_oi.control}
\title{Control parameters in EM algorithm}
\usage{
abun_oi.control(
  epsilon = 1e-05,
  maxit = 5000,
  maxN = NULL,
  maxN_ci = 1e+09,
  start_beta = NULL,
  start_eta = NULL
)
}
\arguments{
\item{epsilon}{positive convergence tolerance \eqn{\epsilon}. The iterations converge when the increase of the log-EL is less than \eqn{\epsilon}.}

\item{maxit}{integer specifying the maximal number of iterations.}

\item{maxN}{number specifying the largest value of \eqn{N} when calculating the estimate of the abundance \eqn{N}. If \code{NULL} the default is 100 times the number of individuals captured.}

\item{maxN_ci}{number specifying the largest value of \eqn{N} when calculating the confidence interval of \eqn{N}.}

\item{start_beta}{vector specifying the starting value of the coefficient \eqn{\beta} in count regression model. If \code{NULL} the default is zero.}

\item{start_eta}{vector specifying the starting value of the coefficient \eqn{\eta} in one-inflated logistic regression model. If \code{NULL} the default is zero.}
}
\value{
A list with the arguments specified.
}
\description{
Various parameters that control performing EM algorithm using \code{\link{abun_oi}}.
}
