% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_oi_class.R
\docType{class}
\name{abun_oi.summary-class}
\alias{abun_oi.summary-class}
\title{abun_oi.summary}
\description{
a S4 class summarizes the point estimates under discrete time capture-recapture models
}
\section{Slots}{

\describe{
\item{\code{coefficients}}{a list, which summarizes the estimates of coefficients in count regression model and one-inflated regression model}

\item{\code{abundance}}{a matrix, which summarizes the estimates of abundance}
}}

