% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_oi_boot.R
\name{abun_oi_boot}
\alias{abun_oi_boot}
\title{Bootstrap-based standard errors and confidence intervals under discrete time CR models}
\usage{
abun_oi_boot(object, boot = 100, seed = 2021)
}
\arguments{
\item{object}{A \code{abun_oi} object.}

\item{boot}{number specifying the number of bootstrap samples. Default is 200.}

\item{seed}{number specifying the random seed. Default is 2021.}
}
\value{
A list with four elements:
        \item{se_N}{the standard error of the abundance estimate.}
        \item{se_beta}{the standard error of the coefficients in count regression model.}
        \item{se_eta}{the standard error of the coefficients in one-inflated regression model.}
        \item{se_alpha}{the standard error of the probability of never being captured.}
        \item{quant_N}{the quantiles of the sampling distribution of the abundance estimate.}
}
\description{
Function to calculate standard errors and confidence intervals based on a bootstrap procedure under discrete time capture-recapture (CR) models.
}
\references{
Liu, Y., Li, P., and Qin, J. (2017).
Maximum empirical likelihood estimation for abundance in a closed population from capture-recapture data.
\emph{Biometrika} \strong{104}, 527-543.
}
