% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_oi_ci.R
\name{abun_oi_ci}
\alias{abun_oi_ci}
\title{EL ratio confidence interval of abundance}
\usage{
abun_oi_ci(object, level = 0.95, maxN_ci = 1e+09)
}
\arguments{
\item{object}{a \code{abun_oi} object.}

\item{level}{number specifying the nominal level.}

\item{maxN_ci}{number specifying the largest searching value.}
}
\value{
a vector, the empirical likelihood ratio confidence interval of abundance.
}
\description{
\code{abun_oi_ci} is used to calculate the empirical likelihood ratio confidence interval of abundance
}
