% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_zt.R
\name{abun_zt}
\alias{abun_zt}
\title{Semiparametric empirical likelihood inference for abundance under zero-truncated models}
\usage{
abun_zt(
  y,
  K = NULL,
  x,
  dist,
  maxN = NULL,
  start_beta = NULL,
  eps = 1e-05,
  maxit = 5000,
  N0 = NULL
)
}
\arguments{
\item{y}{number of times that individuals were captured}

\item{K}{number specifying the number of capture occasions when \code{dist} is \code{"binomial"}.}

\item{x}{vector or matrix containing the individual covariates which have influence on the capture probability.}

\item{dist}{character specification of count regression model family, \code{"poisson"} or \code{"binomial"}.}

\item{maxN}{number specifying the largest searching value for \eqn{N} in EM algorithm.}

\item{start_beta}{vector specifying the starting value of the coefficient \eqn{\beta} in count regression model.}

\item{eps}{positive convergence tolerance \eqn{\epsilon}. The iterations converge in EM algorithm when the increase of the log-EL is less than \eqn{\epsilon}.}

\item{maxit}{integer specifying the maximal number of iterations in EM algorithm.}

\item{N0}{number specifying the value of abundance. If it is \code{NULL}, the maximum EL estimator of abundance is calculated.}
}
\value{
An \code{abun_oi} object.
}
\description{
\code{abun_zt} is used to implement the maximum empirical likelihood method by fitting zero-truncated count regression models.
}
\references{
Liu, Y., Li, P., Liu, Y., and Zhang, R. (2021).
Semiparametric empirical likelihood inference for abundance from one-inflated capture-recapture data.
\emph{Biometrical Journal}.
}
