% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_oi.R
\name{loglikelihood_oi}
\alias{loglikelihood_oi}
\title{Log-EL under the one-inflated count regression model}
\usage{
loglikelihood_oi(N, y, x, beta, z, eta, alpha, prob, dist, K, model)
}
\arguments{
\item{N}{number specifying the population size or abundance.}

\item{y}{vector specifying the number of times that individuals are captured.}

\item{x}{matrix specifying the individual covariates in count regression model.}

\item{beta}{vector specifying the coefficients in count regression model.}

\item{z}{matrix specifying the individual covariates in one-inflated logistic regression model.}

\item{eta}{vector specifying the regression coefficients in one-inflated logistic regression model.}

\item{alpha}{number specifying the probability of never being captured.}

\item{prob}{vector specifying the probability masses of individual covariates.}

\item{dist}{character specification of the count regression model family, \code{"poisson"} (default) or \code{"binomial"}.}

\item{K}{number specifying the number of capture occasions.}

\item{model}{character specification of zero-truncated one-inflated model (\code{"ztoi"}, default) or one-inflated zero-truncated model (\code{"oizt"}).}
}
\value{
A number, the empirical log-likelihood.
}
\description{
Function to calculate the empirical log-likelihood (log-EL) function under the one-inflated count regression model.
}
