% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_zt.R
\name{loglikelihood_zt}
\alias{loglikelihood_zt}
\title{EL log-likelihood under zero-truncated models}
\usage{
loglikelihood_zt(N, y, x, beta, alpha, prob, dist, K)
}
\arguments{
\item{N}{number specifying the value of abundance.}

\item{y}{number of times that individuals were captured}

\item{x}{matrix containing the individual covariates which have influence on the capture probability.}

\item{beta}{vector of regression cofficients in the capture probability model}

\item{alpha}{number, the probability of never being captured}

\item{prob}{vector, the probability mass function of covariates}

\item{dist}{character specification of count regression model family, \code{"poisson"} or \code{"binomial"}.}

\item{K}{number specifying the number of capture occasions when \code{dist} is \code{"binomial"}.}
}
\value{
The value of EL log-likelihood under zero-truncated models.
}
\description{
EL log-likelihood under zero-truncated models
}
\references{
Liu, Y., Li, P., Liu, Y., and Zhang, R. (2021).
Semiparametric empirical likelihood inference for abundance from one-inflated capture-recapture data.
\emph{Biometrical Journal}.
}
