% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_oi_test.R
\name{scoretest_oi}
\alias{scoretest_oi}
\title{Score test for one-inflation}
\usage{
scoretest_oi(object, model = c("ztoi", "oizt"))
}
\arguments{
\item{object}{an abun_oi object produced by \code{\link{abun_oi}} with \code{model} option being \code{"zt"}.}

\item{model}{character specification of zero-truncated one-inflated model (\code{"ztoi"}, default) or one-inflated zero-truncated model (\code{"oizt"}).}
}
\value{
A number, the score test log-likelihood.
}
\description{
Function to calculate the score test statistics for the one-inflation based on zero-truncated one-inflated model and one-inflated zero-truncated model.
}
