% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_oi_class.R
\name{show,abun_oi-method}
\alias{show,abun_oi-method}
\title{Function to show the outcomes of object "abun_oi"}
\usage{
\S4method{show}{abun_oi}(object)
}
\arguments{
\item{object}{an \code{abun_oi} object}
}
\description{
Function to show the outcomes of object "abun_oi"
}
