% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abun_oi_class.R
\name{summary,abun_oi-method}
\alias{summary,abun_oi-method}
\title{Function to summarize the maximum empirical likelihood estimation results}
\usage{
\S4method{summary}{abun_oi}(object, boot = 100, seed = 2021)
}
\arguments{
\item{object}{an \code{abun_oi} object}

\item{boot}{integer specifying the number of bootstrap replications when evaluating the variation.}

\item{seed}{integer specifying the random seed when invoking the bootstrap method \code{\link{abun_oi_boot}}.}
}
\description{
Function to summarize the maximum empirical likelihood estimation results
}
